feat: add hook for supply
