+++ .gitignore
@@ -47,3 +47,5 @@ next-env.d.ts
 !.yarn/plugins
 !.yarn/sdks
 !.yarn/versions
+
+.cursor
